Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    float a;
-    scanf("%f", &a);
-    printf("You entered: %f", a);
+    double a;
+    scanf("%d", &a);
+    printf("You entered: %d", a);
     return 0;
 }
